Removes useRecentLocationUsers
