+++ src/hooks/users/useRecentVenueUsers.ts
@@ -1,9 +1,10 @@
-import { User } from "types/User";
+import { User, UserLocation } from "types/User";
 import { ReactHook } from "types/utility";
 
 import { WithId } from "utils/id";
+import { wrapIntoSlashes } from "utils/string";
 
-import { useRecentLocationUsers } from "hooks/users";
+import { useRecentWorldUsers } from "./useRecentWorldUsers";
 
 export interface UseRecentVenueUsersProps {
   venueId?: string;
@@ -19,12 +20,23 @@ export const useRecentVenueUsers: ReactHook<
   RecentVenueUsersData
 > = ({ venueId }) => {
   const {
-    recentLocationUsers,
-    isRecentLocationUsersLoaded,
-  } = useRecentLocationUsers({ venueId });
+    recentWorldUsers,
+    worldUserLocationsById,
+    isRecentWorldUsersLoaded,
+  } = useRecentWorldUsers();
+
+  const recentLocationUsers = recentWorldUsers.filter((user) => {
+    const userLocation: WithId<UserLocation> | undefined =
+      worldUserLocationsById[user.id];
+
+    return (
+      venueId &&
+      userLocation?.lastVenueIdSeenIn?.includes(wrapIntoSlashes(venueId))
+    );
+  });
 
   return {
     recentVenueUsers: recentLocationUsers as readonly WithId<User>[],
-    isRecentVenueUsersLoaded: isRecentLocationUsersLoaded,
+    isRecentVenueUsersLoaded: isRecentWorldUsersLoaded,
   };
 };
